+++ scripts/systems/SwadeCurrencyCalculator.js
@@ -3,7 +3,7 @@ import CurrencyCalculator from "./CurrencyCalculator.js";
 export default class SwadeCurrencyCalculator extends CurrencyCalculator {
 
     actorCurrency(actor) {
-        return actor.data.data.details.currency;
+        return actor.data.details.currency;
     }
 
     updateActorWithNewFunds(buyer, buyerFunds) {
@@ -39,20 +39,12 @@ export default class SwadeCurrencyCalculator extends CurrencyCalculator {
 
         //console.log("Loot Sheet | Prepare Items");
         // Iterate through items, allocating to containers
-        items = items.sort(function (a, b) {
-            return a.name.localeCompare(b.name);
-        });
-        for (let i of items) {
-            i.img = i.img || DEFAULT_TOKEN;
-            //console.log("Loot Sheet | item", i);
-
-            // Features
-            if (i.type === "weapon") features.weapons.items.push(i);
-            else if (i.type === "armor") features.armor.items.push(i);
-            else if (i.type === "shield") features.shields.items.push(i);
-            else if (i.type === "gear") features.gear.items.push(i);
-            else features.gear.items.push(i);
-        }
+        // for (let i of items) {
+        features.gear.items = items.gear;
+        features.armor.items = items.armor;
+        features.shields.items = items.shield;
+        features.weapons.items = items.weapon;
+        // }
 
         return features;
     }